refactor: simplify ColorPicker and adding Color Popover component
